pyproject.toml
[tool.poetry]
name = "mlops-homework"
version = "0.1.0"
description = ""
authors = ["deout <tolya0906@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.13.0"
pre-commit = "^4.0.1"
isort = "^5.13.2"
loguru = "^0.7.2"
mkdocs = "^1.6.1"
python-dotenv = "^1.0.1"
tqdm = "^4.67.0"
typer = "^0.13.0"
boto3 = "^1.35.56"
s3cmd = "^2.4.0"
pandas = "^2.2.3"
mlflow = "^2.18.0"
scikit-learn = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
